refactor: move hmr plugin logic
