+++ lib/index.js
@@ -11,8 +11,9 @@
 const EventEmitter = require('events');
 
 const Koa = require('koa');
-const { DefinePlugin, HotModuleReplacementPlugin, ProgressPlugin } = require('webpack');
+const { DefinePlugin, ProgressPlugin } = require('webpack');
 
+const { init: initHmrPlugin } = require('./hmr-plugin');
 const { getLogger } = require('./log');
 const { start } = require('./server');
 
@@ -86,61 +87,8 @@ class WebpackPluginServe extends EventEmitter {
       return;
     }
 
-    const chunkFilename = 'wps-hmr.js';
-    const mainFilename = 'wps-hmr.json';
-
     if (this.options.hmr) {
-      const hmrPlugin = new HotModuleReplacementPlugin();
-      const hasHMRPlugin = compiler.options.plugins.some(
-        (plugin) => plugin instanceof HotModuleReplacementPlugin
-      );
-      if (!hasHMRPlugin) {
-        // eslint-disable-next-line no-param-reassign
-        compiler.options.output = Object.assign(compiler.options.output, {
-          hotUpdateChunkFilename: chunkFilename,
-          hotUpdateMainFilename: mainFilename
-        });
-        hmrPlugin.apply(compiler);
-      } else {
-        this.log.warn(
-          'webpack-plugin-serve adds HotModuleReplacementPlugin automatically. Please remove it from your config.'
-        );
-
-        const makePattern = (assetName) =>
-          new RegExp(assetName.replace(/\[\w+\]/g, '(\\w+)').replace(/\./g, '\\.'));
-
-        compiler.hooks.compilation.tap('wps', (compilation) => {
-          compilation.hooks.additionalChunkAssets.intercept({
-            register: (hook) => {
-              if (hook.name === 'HotModuleReplacementPlugin') {
-                const og = hook.fn;
-                // eslint-disable-next-line no-param-reassign
-                hook.fn = (...args) => {
-                  const { assets } = compilation;
-                  const { hotUpdateChunkFilename, hotUpdateMainFilename } = compiler.options.output;
-                  const chunkPattern = makePattern(hotUpdateChunkFilename);
-                  const mainPattern = makePattern(hotUpdateMainFilename);
-
-                  for (const assetName of Object.keys(assets)) {
-                    if (chunkPattern.test(assetName)) {
-                      assets[chunkFilename] = assets[assetName];
-                      delete assets[assetName];
-                    } else if (mainPattern.test(assetName)) {
-                      assets[mainFilename] = assets[assetName];
-                      delete assets[assetName];
-                    }
-                  }
-
-                  const result = og(...args);
-
-                  return result;
-                };
-              }
-              return hook;
-            }
-          });
-        });
-      }
+      initHmrPlugin(compiler);
     }
 
     const { compile, done, invalid, watchClose, watchRun } = compiler.hooks;
